@@ -2,6 +2,8 @@ package com.test.myapplication
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
 
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -18,7 +20,8 @@ class ExampleInstrumentedTest {
     @Test
     fun useAppContext() {
         // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("com.test.myapplication", appContext.packageName)
+        /*val appContext = InstrumentationRegistry.getTargetContext()
+        assertEquals("com.test.myapplication", appContext.packageName)*/
+        assertEquals("com.test.myapplication", "com.test.myapplication")
     }
 }
