@@ -3,9 +3,9 @@ package com.test.myapplication
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import android.view.View
 import android.widget.Button
+import androidx.appcompat.app.AppCompatActivity
 import com.test.myapplication.helpers.GoToUrlHelper
 import com.test.myapplication.providers.MoviesDataProvider
 import com.test.myapplication.providers.MoviesDataProvider.Companion.TYPE_DESCRIPTION
