@@ -3,13 +3,13 @@ package com.test.myapplication
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.fragment.app.Fragment
 import com.test.myapplication.helpers.GoToUrlHelper
 import com.test.myapplication.providers.MoviesDataProvider
 import kotlinx.android.synthetic.main.activity_details_constraint.*
