@@ -1,11 +1,11 @@
 package com.test.myapplication
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.view.ViewPager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.viewpager.widget.ViewPager
 import com.test.myapplication.adapters.DetailsFragmentAdapter
 
 class DetailsGalleryFragment : Fragment() {
